refactor: create a new folder for data_gatherer with abstract class DataSource to implement fetch method for different apis
